% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{caviarpd-package}
\alias{caviarpd-package}
\title{caviarpd: Cluster Analysis via Random Partition Distributions}
\description{
Cluster analysis is performed using pairwise distance information and a random partition distribution. The method is implemented for two random partition distributions. It draws samples and then obtains and plots clustering estimates. An implementation of a selection algorithm is provided for the mass parameter of the partition distribution. Since pairwise distances are the principal input to this procedure, it is most comparable to the hierarchical and k-medoids clustering methods. The method is Dahl, Andros, Carter (2022+) \doi{10.1002/sam.11602}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbdahl/caviarpd-package}
  \item Report bugs at \url{https://github.com/dbdahl/caviarpd-package/issues}
}

}
\author{
\strong{Maintainer}: David B. Dahl \email{dahl@stat.byu.edu} (\href{https://orcid.org/0000-0002-8173-1547}{ORCID})

Authors:
\itemize{
  \item R. Jacob Andros \email{androsrj@gmail.com} (\href{https://orcid.org/0000-0002-1289-385X}{ORCID})
  \item J. Brandon Carter \email{carterj4@icloud.com} (\href{https://orcid.org/0000-0003-1687-0564}{ORCID})
}

Other contributors:
\itemize{
  \item Alex Crichton \email{alex@alexcrichton.com} (Rust crates: cfg-if, proc-macro2) [contributor]
  \item Brendan Zabarauskas \email{bjzaba@yahoo.com.au} (Rust crate: approx) [contributor]
  \item David B. Dahl \email{dahl@stat.byu.edu} (Rust crates: dahl-partition, dahl-salso, epa, roxido, roxido_macro) [contributor]
  \item David Tolnay \email{dtolnay@gmail.com} (Rust crates: proc-macro2, quote, syn, unicode-ident) [contributor]
  \item Jim Turner (Rust crate: ndarray) [contributor]
  \item Jorge Aparicio \email{jorge@japaric.io} (Rust crate: libm) [contributor]
  \item Josh Stone \email{cuviper@gmail.com} (Rust crate: autocfg) [contributor]
  \item Mikhail Vorotilov \email{mikhail.vorotilov@gmail.com} (Rust crate: roots) [contributor]
  \item R. Janis Goldschmidt (Rust crate: matrixmultiply) [contributor]
  \item Sean McArthur \email{sean@seanmonstar.com} (Rust crate: num_cpus) [contributor]
  \item Stefan Lankes (Rust crate: hermit-abi) [contributor]
  \item The Cranelift Project Developers (Rust crate: wasi) [contributor]
  \item The CryptoCorrosion Contributors (Rust crates: ppv-lite86, rand_chacha) [contributor]
  \item The Rand Project Developers (Rust crates: getrandom, rand, rand_chacha, rand_core, rand_distr, rand_pcg) [contributor]
  \item The Rust Project Developers (Rust crates: libc, num-complex, num-integer, num-traits, rand, rand_chacha, rand_core) [contributor]
  \item Ulrik Sverdrup "bluss" (Rust crate: ndarray) [contributor]
  \item bluss (Rust crates: matrixmultiply, rawpointer) [contributor]
}

}
\keyword{internal}
