% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caviarpd.R
\name{caviarpd2}
\alias{caviarpd2}
\title{Cluster Analysis via Random Partition Distributions}
\usage{
caviarpd2(
  distance,
  nClusters,
  nSamples = 500,
  gridLength = 10,
  n0 = 100,
  tol = 0.01,
  loss = "binder",
  temperature = 10,
  similarity = c("exponential", "reciprocal")[1],
  maxNClusters = 0,
  nRuns = 16,
  nCores = 0
)
}
\arguments{
\item{distance}{An object of class 'dist' or a pairwise distance matrix.}

\item{nClusters}{A numeric vector that specifies the range for the number of clusters to consider in the search for a clustering estimate. Should be missing if the \code{mass} argument is used. See \code{Details}.}

\item{nSamples}{The number of samples used to generate the clustering estimate.}

\item{gridLength}{The length of the grid search for an optimal mass parameter. Only applicable if a range of values are provided for \code{nClusters}.}

\item{n0}{An undocumented tuning parameter.}

\item{tol}{An undocumented tuning parameter.}

\item{loss}{The SALSO method (Dahl, Johnson, Müller, 2021) tries to minimize this expected loss when searching the partition space for an optimal estimate. This must be either "binder" or "VI".}

\item{temperature}{A positive number that accentuates or dampens distance between observations.}

\item{similarity}{Either \code{"exponential"} or \code{"reciprocal"} to indicate the desired similarity function.}

\item{maxNClusters}{The maximum number of clusters that can be considered by the SALSO method.}

\item{nRuns}{The number of runs of the SALSO algorithm.}

\item{nCores}{The number of CPU cores to use. A value of zero indicates to use all cores on the system.}
}
\value{
A object of class \code{salso.estimate}, which provides a clustering estimate (a vector of cluster labels) that can be displayed and plotted.
}
\description{
Returns a clustering estimate given pairwise distances using the CaviarPD method.
}
\details{
The \code{mass} argument is the main tuning parameter governing the number of clusters,
with higher values tending toward more clusters. The \code{mass} is a real number bounded
below by 0. When a vector of mass values is supplied, a clustering
estimate for each mass value is generated and the best clustering estimate is returned.

Alternatively, a range for the number of clusters to be considered can be supplied with the
\code{nClusters} argument. Mass values that return a clustering estimate with the minimum and
maximum value of the range will estimated. A grid of mass values (of length \code{gridLength}) between
the estimated min and max cluster mass values will be considered in the search for a clustering
estimate. If \code{nClusters} is a single integer, then a clustering estimate with \code{nClusters}
clusters will be returned.
}
\examples{
# To reduce load on CRAN servers, limit the number of samples, grid length, and CPU cores.
set.seed(34)
iris.dis <- dist(iris[,-5])
est <- caviarpd(distance=iris.dis, mass=c(1, 2), nSamples=50, nCores=1)
samples <- caviarpd(distance=iris.dis, mass=1, nSamples=50, samplesOnly=TRUE, nCores=1)
est <- caviarpd(distance=iris.dis, nClusters=3, nSamples=50, nCores=1)
est <- caviarpd(distance=iris.dis, nClusters=3:5, nSamples=50, gridLength=5, nCores=1)
summ <- summary(est, orderingMethod=2)
plot(summ, type="heatmap")
plot(summ, type="mds")

}
\references{
D. B. Dahl, D. J. Johnson, and P. Müller (2021), Search Algorithms and Loss
Functions for Bayesian Clustering, \href{https://arxiv.org/abs/2105.04451}{arXiv:2105.04451}.
}
