% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloWorld.R
\name{helloWorld}
\alias{helloWorld}
\title{My First Function}
\usage{
helloWorld(name)
}
\arguments{
\item{name}{A character vector giving the names of the people to greet.}
}
\value{
A character vector of the greeting.
}
\description{
This function greets people.
}
