% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Distance for Categorical Attributes}
\usage{
jaccard(data)
}
\arguments{
\item{data}{The data set of categorical variables for which the jaccard distance is computed.}
}
\value{
An object of class "dist", which can easily be converted to an symmetric \eqn{n*n}
matrix. See the dist() function documentation for more details.
}
\description{
Computes the pairwise distance matrix for categorical variables.
}
\examples{
jaccard(npk[,-5])
jaccard(warpbreaks[,-1])

}
