% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss.indexes}
\alias{loss.indexes}
\title{Accuracy for a List of Clustering Estimates through Loss Functions}
\usage{
loss.indexes(estimate.list, truth, order.by = "binder")
}
\arguments{
\item{estimate.list}{An object of type 'list' containing any number of clustering estimates. Estimates can be from hierarchical clustering, caviarPD, or another method, but must be in list form.}

\item{truth}{The true partition of the data against which each estimate is compared, in numeric form.}

\item{order.by}{Since loss is computed for each estimate under both binder and VI, this argument specifies which loss function values the estimates are sorted by in the returned data frame.}
}
\value{
A data.frame object which contains each estimate's name (from the original list), number of clusters, VI loss, and Binder loss.
}
\description{
Computes VI and Binder loss for any number of given clustering estimates. Estimates must be passed to function in list form.
}
\examples{
iris.dis <- dist(iris[,-5])
iris.truth <- as.numeric(iris[,5])
est1 <- caviarPD(distance=iris.dis, mass=0.9, nSamples=1000, loss='binder')
est2 <- caviarPD(distance=iris.dis, mass=2.0, loss="VI")
loss.indexes(list(est1, est2), iris.truth)

}
