% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masses.R
\name{select.masses}
\alias{select.masses}
\title{Mass Parameter Selection for the CaviarPD Procedure}
\usage{
select.masses(
  dis,
  ncl.range,
  single = FALSE,
  nSD = 3,
  discount = 0,
  temperature = 10,
  loss = "binder",
  nSamplesA = 500,
  nSamplesB = 1000,
  w = c(1, 1, 0)
)
}
\arguments{
\item{dis}{Pairwise distance matrix of class 'dist'.}

\item{ncl.range}{A vector of two values representing the minimum and maximum number of clusters that the user wishes to consider in selecting mass values.}

\item{single}{If TRUE, the algorithm returns both a list of masses for each targeted cluster count as well as a best overal mass selected from that list.}

\item{nSD}{Number of standard deviations in the EPA distribution to consider for a range of mass values; lower values allow the algorithm to run more quickly but risk an error if deemed insufficient.}

\item{discount}{Typically 0, controls the distribution of subset sizes.}

\item{temperature}{A positive number that accentuates or dampens distance between observations.}

\item{loss}{The salso method aims to estimate this loss function when searching the partition space for an optimal estimate, must be specified as either "binder" or "VI".}

\item{nSamplesA}{Number of samples used by the salso method to repeatedly calculate the cluster count for each attempted mass value.}

\item{nSamplesB}{Number of samples used by the salso method to obtain the optimal mass; only applicable if single=TRUE.}

\item{w}{Weights for selecting a single mass. The first weight is attached to the partition confidence, the second weight is attached to the variance trio, and the last is attached to the nunber of clusters.}
}
\value{
If single==FALSE, returns a list with two elements: a chain of mass values corresponding to each cluster count,
and the mass value that had the best overall confidence plot. If single==TRUE, returns only the best overall mass value.
}
\description{
User inputs a range of clustering sizes to obtain a mass value that can viably correspond to each cluster count.
}
\examples{
tooth.dis <- dist(scale(ToothGrowth[,-2]))
# In practice the user should do at least 100 runs, but for ease of testing we use less here.
select.masses(tooth.dis, ncl.range=c(2,4), nSamplesA=10, nSamplesB=10)
iris.dis <- dist(iris[,-5])
select.masses(iris.dis, ncl.range=c(3,6), single=TRUE, nSamplesA=10, nSamplesB=10)

}
