% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masses.R
\name{single.mass}
\alias{single.mass}
\title{Single Mass Parameter Selection for the CaviarPD Procedure}
\usage{
single.mass(
  masses,
  distance,
  temperature = 10,
  discount = 0,
  nSamples = 1000,
  w = c(1, 1, 0),
  loss = "binder"
)
}
\arguments{
\item{masses}{A vector of mass values from which to select the best mass. This can be a simple sequence for some range or a list of masses generated from the select.masses function.}

\item{distance}{Pairwise distance matrix of class 'dist'.}

\item{temperature}{A positive number that accentuates or dampens distance between observations.}

\item{discount}{Typically 0, controls the distribution of subset sizes.}

\item{nSamples}{The number of samples used to estimate the loss function in the salso method.}

\item{w}{Weights for selecting a single mass. The first weight is attached to the partition confidence, the second weight is attached to the variance trio, and the last is attached to the nunber of clusters.}

\item{loss}{The salso method aims to estimate this loss function when searching the partition space for an optimal estimate, must be specified as either "binder" or "VI".}
}
\value{
The value of the mass parameter that should be used in the caviarPD method for the given pairwise distance matrix.
}
\description{
Calculates the partition confidence and variance ratios for each mass value to find the best mass. User can input masses from the select.masses function or supply their own.
}
\examples{
iris.dis <- dist(iris[,-5])
iris.masses <- select.masses(iris.dis, ncl.range=c(3,6), nSamplesA=10, nSamplesB=10)
single.mass(masses=iris.masses, distance=iris.dis, nSamples=10)
single.mass(masses=seq(.5, 2, by=.25), distance=iris.dis, nSamples=10)

}
