% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caviarPD.R
\name{caviarpd}
\alias{caviarpd}
\title{Cluster Analysis via Random Partition Distributions}
\usage{
caviarpd(
  distance,
  mass = 1,
  loss = "binder",
  nSamples = 1000,
  samplesOnly = FALSE,
  distr = "EPA",
  temperature = 10,
  discount = 0,
  maxNClusters = 0,
  nCores = 0
)
}
\arguments{
\item{distance}{A pairwise distance matrix of class 'dist'.}

\item{mass}{The main tuning parameter for the number of clusters. A higher mass value tends to lead to more clusters.}

\item{loss}{The SALSO method (Dahl, Johnson, Müller, 2021) tries to minimize this expected loss when searching the partition space for an optimal estimate. This must be either "binder" or "VI".}

\item{nSamples}{Number of samples used to estimate the partition confidence and variance ratio for each mass value obtained from the search algorithm; only applicable if single=TRUE.}

\item{samplesOnly}{If TRUE, the function only returns the samples generated for a given mass, temperature, and discount rather than an actual clustering estimate.}

\item{distr}{The random partition distribution used to generate samples.  This must be specified as either "EPA" or "ddCRP".}

\item{temperature}{A positive number that accentuates or dampens distance between observations.}

\item{discount}{A number in \eqn{[0,1)} giving the discount parameter to control the distribution of subset sizes.}

\item{maxNClusters}{The maximum number of clusters that can be considered by the SALSO method.}

\item{nCores}{The number of CPU cores to use. A value of zero indicates to use all cores on the system.}
}
\value{
A list containing two elements: the clustering estimate and the summary. The summary contains the pairwise probabilities for all samples and is used to create the confidence plots.
}
\description{
Returns a clustering estimate given pairwise distances. Users can also specify parameters for the mass, discount, temperature, loss function, or number of samples.
}
\examples{
iris.dis <- dist(iris[,-5])
# In practice the user should use at least 100 samples, but for ease of testing we use less here.
caviarpd(distance=iris.dis, nSamples=10, nCores=1)
caviarpd(distance=iris.dis, mass=0.75, loss="binder", nSamples=10, maxNClusters=3, nCores=1)

}
\references{
D. B. Dahl, D. J. Johnson, and P. Müller (2021), Search Algorithms and Loss
Functions for Bayesian Clustering, \href{https://arxiv.org/abs/2105.04451}{arXiv:2105.04451}.
}
