% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caviarpd.R
\name{caviarpd}
\alias{caviarpd}
\title{Cluster Analysis via Random Partition Distributions}
\usage{
caviarpd(
  distance,
  nClusters,
  mass,
  nSamples = 1000,
  gridLength = 10,
  samplesOnly = FALSE,
  loss = "binder",
  distr = "EPA",
  temperature = 10,
  discount = 0,
  sd = 3,
  maxNClusters = 0,
  nCores = 0
)
}
\arguments{
\item{distance}{A pairwise distance matrix of class 'dist'.}

\item{nClusters}{A numeric vector giving the range for the number of clusters to consider when search for the mass parameter. Should be unset if the \code{mass} argument is used.}

\item{mass}{The main tuning parameter governing the number of clusters, with higher values tends toward more clusters. Should be unset if the \code{nClusters} argument is provided.}

\item{nSamples}{The number of samples used to generate the clustering estimate.}

\item{gridLength}{The length of the grid search for an optimal mass parameter. Only applicable if multiple values are provided for \code{nClusters}.}

\item{samplesOnly}{If TRUE, the function only returns the samples generated for a given mass, temperature, and discount rather than an actual clustering estimate.}

\item{loss}{The SALSO method (Dahl, Johnson, Müller, 2021) tries to minimize this expected loss when searching the partition space for an optimal estimate. This must be either "binder" or "VI".}

\item{distr}{The random partition distribution used to generate samples.  This must be specified as either "EPA" or "ddCRP".}

\item{temperature}{A positive number that accentuates or dampens distance between observations.}

\item{discount}{A number in \eqn{[0,1)} giving the discount parameter to control the distribution of subset sizes.}

\item{sd}{Number of standard deviations away from the expectation to be considered in finding boundary mass values.}

\item{maxNClusters}{The maximum number of clusters that can be considered by the SALSO method.}

\item{nCores}{The number of CPU cores to use. A value of zero indicates to use all cores on the system.}
}
\value{
A clustering estimate. The summary of the estimate contains the pairwise probabilities for all samples and is used to create the confidence plots.
}
\description{
Returns a clustering estimate given pairwise distances. Users can also specify parameters for the mass, discount, temperature, loss function, or number of samples.
}
\examples{
iris.dis <- dist(iris[,-5])
# In practice the user should use at least 100 samples, but for ease of testing we use less here.
caviarpd(distance=iris.dis, nClusters=2:6, nSamples=10, nCores=1)
caviarpd(distance=iris.dis, mass=1, nSamples=10, nCores=1)

}
\references{
D. B. Dahl, D. J. Johnson, and P. Müller (2021), Search Algorithms and Loss
Functions for Bayesian Clustering, \href{https://arxiv.org/abs/2105.04451}{arXiv:2105.04451}.
}
